5_TuningPrompting/requirements.txt
pandas
openai
openpyxl
streamlit
dspy
aisuite[openai,google,groq]
google.generativeai